requirements.txt
django>=1.4.16,<1.5
django-compat